refactor: extract asset type determination logic into util module
