+++ lib/dependencies/URLDependency.js
@@ -10,6 +10,7 @@ const RawDataUrlModule = require("../asset/RawDataUrlModule");
 const {
 	getDependencyUsedByExportsCondition
 } = require("../optimize/InnerGraph");
+const getAssetType = require("../util/assetType");
 const makeSerializable = require("../util/makeSerializable");
 const memoize = require("../util/memoize");
 const ModuleDependency = require("./ModuleDependency");
@@ -28,34 +29,6 @@ const ModuleDependency = require("./ModuleDependency");
 const getIgnoredRawDataUrlModule = memoize(
 	() => new RawDataUrlModule("data:,", "ignored-asset", "(ignored asset)")
 );
-
-/**
- * Determines the 'as' attribute value for prefetch/preload based on file extension
- * @param {string} request The module request string
- * @returns {string} The 'as' attribute value
- */
-const getAssetType = (request) => {
-	// Reference: MDN rel=preload documentation (https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Attributes/rel/preload#what_types_of_content_can_be_preloaded)
-	// Valid 'as' values: fetch, font, image, script, style, track
-	if (/\.(png|jpe?g|gif|svg|webp|avif|bmp|ico|tiff?)$/i.test(request)) {
-		return "image";
-	} else if (/\.(woff2?|ttf|otf|eot)$/i.test(request)) {
-		return "font";
-	} else if (/\.(js|mjs|jsx|ts|tsx)$/i.test(request)) {
-		return "script";
-	} else if (/\.css$/i.test(request)) {
-		return "style";
-	} else if (/\.vtt$/i.test(request)) {
-		return "track"; // WebVTT files for video subtitles/captions
-	} else if (
-		/\.(mp4|webm|ogg|mp3|wav|flac|aac|m4a|avi|mov|wmv|mkv)$/i.test(request)
-	) {
-		// Audio/video files: use 'fetch' as fallback since as='audio'/'video' not supported
-		// Reference: https://github.com/mdn/browser-compat-data/issues/9577
-		return "fetch";
-	}
-	return "fetch"; // Generic fetch for unknown types
-};
 
 class URLDependency extends ModuleDependency {
 	/**
