+++ lib/prefetch/AssetPrefetchStartupPlugin.js
@@ -6,6 +6,7 @@
 "use strict";
 
 const RuntimeGlobals = require("../RuntimeGlobals");
+const getAssetType = require("../util/assetType");
 const AssetPrefetchStartupRuntimeModule = require("./AssetPrefetchStartupRuntimeModule");
 
 /** @typedef {import("../Chunk")} Chunk */
@@ -105,8 +106,7 @@ class AssetPrefetchStartupPlugin {
 									assetUrl = request.split(/[\\/]/).pop() || request;
 								}
 
-								const assetType =
-									AssetPrefetchStartupPlugin._getAssetType(request);
+								const assetType = getAssetType(request);
 								const info = {
 									url: assetUrl,
 									as: assetType,
@@ -175,30 +175,6 @@ class AssetPrefetchStartupPlugin {
 					set.add(RuntimeGlobals.publicPath);
 				});
 		});
-	}
-
-	/**
-	 * Determines the 'as' attribute value for prefetch/preload based on file extension
-	 * @param {string} request The module request string
-	 * @returns {string} The 'as' attribute value
-	 */
-	static _getAssetType(request) {
-		if (/\.(png|jpe?g|gif|svg|webp|avif|bmp|ico|tiff?)$/i.test(request)) {
-			return "image";
-		} else if (/\.(woff2?|ttf|otf|eot)$/i.test(request)) {
-			return "font";
-		} else if (/\.(js|mjs|jsx|ts|tsx)$/i.test(request)) {
-			return "script";
-		} else if (/\.css$/i.test(request)) {
-			return "style";
-		} else if (/\.vtt$/i.test(request)) {
-			return "track";
-		} else if (
-			/\.(mp4|webm|ogg|mp3|wav|flac|aac|m4a|avi|mov|wmv|mkv)$/i.test(request)
-		) {
-			return "fetch";
-		}
-		return "fetch";
 	}
 }
